add comment

+++ G64_Lab5/drivers/inc/audio_driver.h
@@ -1,7 +1,6 @@
 #ifndef __AUDIO_DRIVER
 #define __AUDIO_DRIVER
 
-	extern int audio_write_ASM(float data);
-//	extern int audio_write_ASM(int data);
+	extern int audio_write_ASM(int data);
 
 #endif
